move changes from pyth-js
